feat : 화면에 적용되지않던 오류 수정
