+++ src/main/java/roomescape/controller/ReservationController.java
@@ -5,14 +5,13 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import roomescape.domain.Reservation;
 import roomescape.dto.ReservationDTO;
-import roomescape.dto.ReservationResponseDTO;
-
 import roomescape.service.ReservationService;
 
 import java.net.URI;
 import java.util.List;
 
 @Controller
+@RequestMapping("/reservations")
 public class ReservationController {
     private final ReservationService reservationService;
 
@@ -20,27 +19,22 @@ public class ReservationController {
         this.reservationService = reservationService;
     }
 
-    @GetMapping("/reservation")
-    public String getReservationPage() {
-        return "new-reservation";
-    }
-
-    @GetMapping("/reservations")
+    @GetMapping
     public ResponseEntity<List<Reservation>> reservations() {
         List<Reservation> reservations = reservationService.findAllReservations();
 
         return ResponseEntity.ok().body(reservations);
     }
 
-    @PostMapping("/reservations")
-    public ResponseEntity<ReservationResponseDTO> makeReservation(@RequestBody ReservationDTO reservationDTO) {
+    @PostMapping
+    public ResponseEntity<Reservation> makeReservation(@RequestBody ReservationDTO reservationDTO) {
 
-        ReservationResponseDTO newReservation = reservationService.makeReservation(reservationDTO);
+        Reservation newReservation = reservationService.makeReservation(reservationDTO);
 
-        return ResponseEntity.created(URI.create("/reservations/" + newReservation.id())).body(newReservation);
+        return ResponseEntity.created(URI.create("/reservations/" + newReservation.getId())).body(newReservation);
     }
 
-    @DeleteMapping("/reservations/{id}")
+    @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteReservation(@PathVariable("id") Long id) {
         reservationService.deleteReservation(id);
